requirements.txt
aiopg
aiocryptocurrency==0.1.8
aiofiles
aioredis
aiohttp
bcrypt
beautifulsoup4
bleach==4.1.0
timeago
captcha
email-validator
image
peewee
peewee-async
python-dateutil
pyqrcode
pypng
pygments
jinja2==3.0.3
mistletoe
quart
quart_session
quart_schema
quart_session_openid
python-slugify
